+++ src/scripts/components/index.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import MUIThemeTester from './test'
+import MUIThemeTester from 'test/mui-theme-tester'
 
 const Application = () => {
     return (
